+++ client/routes.jsx
@@ -3,6 +3,7 @@ import React from 'react';
 import App from './modules/App/App';
 import IDEView from './modules/IDE/pages/IDEView';
 import FullView from './modules/IDE/pages/FullView';
+import LoginView from './modules/User/pages/LoginView';
 // import SketchListView from './modules/Sketch/pages/SketchListView';
 import { getUser } from './modules/User/actions';
 
@@ -14,7 +15,7 @@ const routes = (store) =>
   (
   <Route path="/" component={App}>
     <IndexRoute component={IDEView} onEnter={checkAuth(store)} />
-    <Route path="/login" component={IDEView} />
+    <Route path="/login" component={LoginView} />
     <Route path="/signup" component={IDEView} />
     <Route path="/reset-password" component={IDEView} />
     <Route path="/reset-password/:reset_password_token" component={IDEView} />
