+++ client/styles/components/_login.scss
@@ -1,10 +1,9 @@
 .login {
-	@extend %modal;
 	text-align: center;
+	height: 100%;
 	display: flex;
 	flex-direction: column;
 	justify-content: center;
-	padding: #{20 / $base-font-size}rem;
   align-items: center;
 }
